Repasando las rutas. Voy por los estilos del recuperar cuenta.blade. Con los mensajes de validación

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\ApiAuthentication;
 use App\Http\Controllers\ArticuloController;
 use App\Http\Controllers\ClaseController;
 use App\Http\Controllers\EjercicioController;
@@ -22,39 +23,40 @@ use Illuminate\Support\Facades\Route;
 /////// RUTAS DE AUTENTICACIÓN ///////
 //////////////////////////////////////
 
-Route::post("/iniciar-sesion",
+Route::post("/iniciar-sesion", //
     [ApiAuthentication::class, "login"]
 )->middleware("guest");
 
 Route::post("/registrarse",
     [ApiAuthentication::class, "register"]
-)->middleware("guest");
+)
+    ->middleware("guest");
 
 Route::post("/recuperar-cuenta",
     [ApiAuthentication::class, "recuperarCuenta"]
 );
 
-Route::get("/verificar-cuenta",
+Route::get("/verificar-cuenta",//
     [ApiAuthentication::class, "mandarCorreoVerificacionCuenta"]
 )->middleware("auth:sanctum");
 
-Route::post("/cambiar-contrasena",
+Route::post("/cambiar-contrasena",//
     [ApiAuthentication::class, "cambiarContrasena"]
 )->middleware("auth:sanctum", "cuentaVerificada");
 
-Route::post("/ajustes-cuenta",
+Route::post("/ajustes-cuenta",//
     [UserController::class, "guardarAjustesCuentaUsuario"]
 )->middleware("auth:sanctum", "cuentaVerificada");
 
-Route::get("/ajustes-cuenta",
+Route::get("/ajustes-cuenta",//
     [UserController::class, "leerAjustesCuentaUsuario"]
 )->middleware("auth:sanctum", "cuentaVerificada");
 
-Route::get("/eliminar-cuenta",
+Route::get("/eliminar-cuenta",//
     [UserController::class, "eliminarCuenta"]
 )->middleware("auth:sanctum", "cuentaVerificada");
 
-Route::post("/enviar-sugerencia",
+Route::post("/enviar-sugerencia",//
     [UserController::class, "enviarSugerencia"]
 )->middleware("auth:sanctum", "cuentaVerificada");
 
@@ -62,7 +64,7 @@ Route::post("/enviar-sugerencia",
 /////// RUTAS DE LOG DE ERRORES ///////
 ///////////////////////////////////////
 
-Route::post("/log-error", function(Request $request){
+Route::post("/log-error", function(Request $request){//
     Log::channel("front")->error($request->get("message"), $request->get("context"));
 }
 );
@@ -76,7 +78,7 @@ Route::post("/log-error", function(Request $request){
 ///// USUARIO /////
 ///////////////////
 
-Route::get("iniciar-sesion", function() {
+Route::get("iniciar-sesion", function() {//
     $user = User::find(1);
     $token = $user->createToken("authToken")->plainTextToken;
 
@@ -89,48 +91,48 @@ Route::get("iniciar-sesion", function() {
 ///// GIMNASIOS /////
 /////////////////////
 
-Route::get("gimnasios",
+Route::get("gimnasios",//
     [GimnasioController::class, "misGimnasios"])
     ->middleware(["auth:sanctum", "cuentaVerificada"])
     ->name("mis-gimnasios");
 
-Route::post("gimnasios",
+Route::post("gimnasios",//
     [GimnasioController::class, "crearGimnasio"])
     ->middleware(["auth:sanctum", "cuentaVerificada"])
     ->name("crear-gimnasio");
 
-Route::put("gimnasios/{gimnasio}",
+Route::put("gimnasios/{gimnasio}",//
     [GimnasioController::class, "editarGimnasio"])
     ->middleware(["auth:sanctum", "cuentaVerificada"])
     ->can("editarGimnasio", "gimnasio")
     ->name("editar-gimnasio");
 
-Route::delete("gimnasios/{gimnasio}",
+Route::delete("gimnasios/{gimnasio}",//
     [GimnasioController::class, "eliminarGimnasio"])
     ->middleware(["auth:sanctum", "cuentaVerificada"])
     ->can("eliminarGimnasio", "gimnasio")
     ->name("eliminar-gimnasio");
 
-Route::post("gimnasios/{gimnasio}/invitar-usuario",
+Route::post("gimnasios/{gimnasio}/invitar-usuario",//
     [GimnasioController::class, "invitarUsuario"])
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("invitarUsuarios", "gimnasio")
     ->name("invitar-usuario");
 
-Route::get("gimnasios/{gimnasio}/reenviar-invitacion/{usuario}",
+Route::get("gimnasios/{gimnasio}/reenviar-invitacion/{usuario}",//
     [GimnasioController::class, "reenviarInvitacion"])
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("reenviarInvitaciones", "gimnasio")
     ->name("reenviar-invitacion");
 
-Route::get("gimnasios/{gimnasio}/crear-administrador/{usuario}",
+Route::get("gimnasios/{gimnasio}/crear-administrador/{usuario}",//
     [GimnasioController::class, "anyadirAdministrador"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearAdministradores", [Gimnasio::class, "gimnasio", "usuario"])
     ->name("crear-administrador");
 
-Route::delete("gimnasios/{gimnasio}/quitar-administrador/{usuario}",
+Route::delete("gimnasios/{gimnasio}/quitar-administrador/{usuario}",//
     [GimnasioController::class, "quitarAdministrador"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -141,40 +143,40 @@ Route::delete("gimnasios/{gimnasio}/quitar-administrador/{usuario}",
 ///// CLASES /////
 //////////////////
 
-Route::get("gimnasios/{gimnasio}/clases",
+Route::get("gimnasios/{gimnasio}/clases",//
     [ClaseController::class, "verClases"])
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verClases", [Clase::class, "gimnasio"])
     ->name("ver-clases-de-gimnasio");
 
-Route::post("gimnasios/{gimnasio}/clases",
+Route::post("gimnasios/{gimnasio}/clases",//
     [ClaseController::class, "crearClase"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearClases", [Clase::class, "gimnasio"])
     ->name("crear-clase");
 
-Route::put("gimnasios/{gimnasio}/clases/{clase}",
+Route::put("gimnasios/{gimnasio}/clases/{clase}",//
     [ClaseController::class, "editarClase"]
 )->middleware("auth:sanctum", "cuentaVerificada")
     ->can("editarClases", [Clase::class, "gimnasio", "clase"])
     ->name("editar-clase");
 
-Route::delete("gimnasios/{gimnasio}/clases/{clase}",
+Route::delete("gimnasios/{gimnasio}/clases/{clase}",//
     [ClaseController::class, "eliminarClase"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("eliminarClases", [Clase::class, "gimnasio", "clase"])
     ->name("eliminar-clase");
 
-Route::get("gimnasios/{gimnasio}/clases/{clase}/apuntarse",
+Route::get("gimnasios/{gimnasio}/clases/{clase}/apuntarse",//
     [ClaseController::class, "usuarioSeApunta"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("usuarioSePuedeApuntar", [Clase::class, "gimnasio", "clase"])
     ->name("usuario-se-apunta");
 
-Route::get("gimnasios/{gimnasio}/clases/{clase}/desapuntarse",
+Route::get("gimnasios/{gimnasio}/clases/{clase}/desapuntarse",//
     [ClaseController::class, "usuarioSeDesapunta"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -187,28 +189,28 @@ Route::get("gimnasios/{gimnasio}/clases/{clase}/desapuntarse",
 ///// TARIFAS /////
 ///////////////////
 
-Route::get("gimnasios/{gimnasio}/tarifas",
+Route::get("gimnasios/{gimnasio}/tarifas",//
     [TarifaController::class, "verTarifas"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verTarifas", [Tarifa::class, "gimnasio"])
     ->name("ver-tarifas");
 
-Route::post("gimnasios/{gimnasio}/tarifas",
+Route::post("gimnasios/{gimnasio}/tarifas",//
     [TarifaController::class, "crearTarifa"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearTarifas", [Tarifa::class, "gimnasio"])
     ->name("crear-tarifas");
 
-Route::put("gimnasios/{gimnasio}/tarifas/{tarifa}",
+Route::put("gimnasios/{gimnasio}/tarifas/{tarifa}",//
     [TarifaController::class, "modificarTarifa"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("editarTarifas", [Tarifa::class, "gimnasio", "tarifa"])
     ->name("editar-tarifas");
 
-Route::delete("gimnasios/{gimnasio}/tarifas/{tarifa}",
+Route::delete("gimnasios/{gimnasio}/tarifas/{tarifa}",//
     [TarifaController::class, "eliminarTarifa"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -221,7 +223,7 @@ Route::delete("gimnasios/{gimnasio}/tarifas/{tarifa}",
 ///// SUSCRIPCIONES /////
 /////////////////////////
 
-Route::get("gimnasios/{gimnasio}/suscripciones",
+Route::get("gimnasios/{gimnasio}/suscripciones",//
     [SuscripcionController::class, "verSuscripciones"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -229,7 +231,7 @@ Route::get("gimnasios/{gimnasio}/suscripciones",
     ->name("ver-suscripcion");
 
 //Generar una suscripción como user
-Route::post("gimnasios/{gimnasio}/suscripciones",
+Route::post("gimnasios/{gimnasio}/suscripciones",//
     [SuscripcionController::class, "crearSuscripcion"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -237,28 +239,28 @@ Route::post("gimnasios/{gimnasio}/suscripciones",
     ->name("crear-suscripcion");
 
 //Generar suscripción como admin
-Route::post("gimnasios/{gimnasio}/suscribirse",
+Route::post("gimnasios/{gimnasio}/suscribirse",//
     [SuscripcionController::class, "adminCreaSuscripcion"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearSuscripcionesComoAdmin", [Suscripcion::class, "gimnasio"])
     ->name("admin-crear-suscripcion");
 
-Route::put("gimnasios/{gimnasio}/suscripciones/{suscripcion}",
+Route::put("gimnasios/{gimnasio}/suscripciones/{suscripcion}",//
     [SuscripcionController::class, "editarSuscripcion"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("editarSuscripciones", [Suscripcion::class, "gimnasio", "suscripcion"])
     ->name("editar-suscripcion");
 
-Route::delete("gimnasios/{gimnasio}/suscripciones/{suscripcion}",
+Route::delete("gimnasios/{gimnasio}/suscripciones/{suscripcion}",//
     [SuscripcionController::class, "eliminarSuscripcion"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("eliminarSuscripciones", [Suscripcion::class, "gimnasio", "suscripcion"])
     ->name("eliminar-suscripcion");
 
-Route::get("gimnasios/{gimnasio}/suscripciones/{suscripcion}/marcar-pagada",
+Route::get("gimnasios/{gimnasio}/suscripciones/{suscripcion}/marcar-pagada",//
     [SuscripcionController::class, "marcarSuscripcionComoPagada"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -271,49 +273,49 @@ Route::get("gimnasios/{gimnasio}/suscripciones/{suscripcion}/marcar-pagada",
 ///// ARTÍCULOS /////
 /////////////////////
 
-Route::get("gimnasios/{gimnasio}/articulos",
+Route::get("gimnasios/{gimnasio}/articulos",//
     [ArticuloController::class, "verArticulos"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verArticulos", [Articulo::class, "gimnasio"])
     ->name("ver-articulos");
 
-Route::post("gimnasios/{gimnasio}/articulos",
+Route::post("gimnasios/{gimnasio}/articulos",//
     [ArticuloController::class, "crearArticulo"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearArticulos", [Articulo::class, "gimnasio"])
     ->name("crear-articulo");
 
-Route::put("gimnasios/{gimnasio}/articulos/{articulo}",
+Route::put("gimnasios/{gimnasio}/articulos/{articulo}",//
     [ArticuloController::class, "editarArticulo"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("editarArticulos", [Articulo::class, "gimnasio", "articulo"])
     ->name("editar-articulo");
 
-Route::delete("gimnasios/{gimnasio}/articulos/{articulo}",
+Route::delete("gimnasios/{gimnasio}/articulos/{articulo}",//
     [ArticuloController::class, "eliminarArticulo"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("eliminarArticulos", [Articulo::class, "gimnasio", "articulo"])
     ->name("eliminar-articulo");
 
-Route::get("gimnasios/{gimnasio}/articulos/historial-compras",
+Route::get("gimnasios/{gimnasio}/articulos/historial-compras",//
     [ArticuloController::class, "historialDeCompras"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verMiHistorialDeCompras", [Articulo::class, "gimnasio"])
     ->name("articulos-historial-compras");
 
-Route::get("gimnasios/{gimnasio}/articulos/{articulo}/comprar",
+Route::get("gimnasios/{gimnasio}/articulos/{articulo}/comprar",//
     [ArticuloController::class, "comprarArticulo"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("comprarArticulos", [Articulo::class, "gimnasio", "articulo"])
     ->name("comprar-articulo");
 ;
-Route::get("gimnasios/{gimnasio}/articulos/pagar-compra/{compra}",
+Route::get("gimnasios/{gimnasio}/articulos/pagar-compra/{compra}",//
     [ArticuloController::class, "pagarCompra"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -327,42 +329,42 @@ Route::get("gimnasios/{gimnasio}/articulos/pagar-compra/{compra}",
 ///// EJERCICIOS /////
 //////////////////////
 
-Route::get("gimnasios/{gimnasio}/ejercicios",
+Route::get("gimnasios/{gimnasio}/ejercicios",//
     [EjercicioController::class, "verEjercicios"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verEjercicios", [Ejercicio::class, "gimnasio"])
     ->name("ver-ejercicios");
 
-Route::post("gimnasios/{gimnasio}/ejercicios",
+Route::post("gimnasios/{gimnasio}/ejercicios",//
     [EjercicioController::class, "crearEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearEjercicios", [Ejercicio::class, "gimnasio"])
     ->name("crear-ejercicio");
 
-Route::put("gimnasios/{gimnasio}/ejercicios/{ejercicio}",
+Route::put("gimnasios/{gimnasio}/ejercicios/{ejercicio}",//
     [EjercicioController::class, "modificarEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("modificarEjercicios", [Ejercicio::class, "gimnasio", "ejercicio"])
     ->name("modificar-ejercicio");
 
-Route::delete("gimnasios/{gimnasio}/ejercicios/{ejercicio}",
+Route::delete("gimnasios/{gimnasio}/ejercicios/{ejercicio}",//
     [EjercicioController::class, "eliminarEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("eliminarEjercicios", [Ejercicio::class, "gimnasio", "ejercicio"])
     ->name("eliminar-ejercicio");
 
-Route::get("gimnasios/{gimnasio}/clases/{clase}/asignar-ejercicio/{ejercicio}",
+Route::get("gimnasios/{gimnasio}/clases/{clase}/asignar-ejercicio/{ejercicio}",//
     [EjercicioController::class, "asociarEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("asociarEjerciciosAClase", [Ejercicio::class, "gimnasio", "clase", "ejercicio"])
     ->name("asociar-ejercicio");
 
-Route::get("gimnasios/{gimnasio}/clases/{clase}/desasignar-ejercicio/{ejercicio}",
+Route::get("gimnasios/{gimnasio}/clases/{clase}/desasignar-ejercicio/{ejercicio}",//
     [EjercicioController::class, "desasociarEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -376,19 +378,19 @@ Route::get("gimnasios/{gimnasio}/clases/{clase}/desasignar-ejercicio/{ejercicio}
 ///// MÉTRICAS /////
 ////////////////////
 
-Route::get("metricas",
+Route::get("metricas",//
     [MetricaController::class, "verMetricas"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->name("ver-metricas");
 
-Route::post("metricas",
+Route::post("metricas",//
     [MetricaController::class, "crearMetrica"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->name("crear-metrica");
 
-Route::delete("metricas/{metrica}",
+Route::delete("metricas/{metrica}",//
     [MetricaController::class, "eliminarMetrica"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
@@ -403,28 +405,28 @@ Route::delete("metricas/{metrica}",
 ///// EJERCICIOUSUARIO /////
 ////////////////////////////
 
-Route::get("gimnasios/{gimnasio}/registros-de-peso",
+Route::get("gimnasios/{gimnasio}/registros-de-peso",//
     [EjerciciosUsuariosController::class, "verRegistrosDePeso"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verRegistrosDePeso", [EjercicioUsuario::class, "gimnasio"])
     ->name("ver-registros-de-peso");
 
-Route::get("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso",
+Route::get("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso",//
     [EjerciciosUsuariosController::class, "verRegistrosDePesoPorEjercicio"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("verRegistrosDePesoPorEjercicio", [EjercicioUsuario::class, "gimnasio", "ejercicio"])
     ->name("ver-registros-de-peso-por-ejercicio");
 
-Route::post("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso",
+Route::post("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso",//
     [EjerciciosUsuariosController::class, "registrarNuevaMarcaDePeso"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
     ->can("crearRegistrosDePeso", [EjercicioUsuario::class, "gimnasio", "ejercicio"])
     ->name("crear-registros-de-peso");
 
-Route::delete("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso/{ejercicioUsuario}",
+Route::delete("gimnasios/{gimnasio}/ejercicios/{ejercicio}/registros-de-peso/{ejercicioUsuario}",//
     [EjerciciosUsuariosController::class, "eliminarMarcaDePeso"]
 )
     ->middleware("auth:sanctum", "cuentaVerificada")
